MD-6, MD-9 refactoring
