+++ app/src/main/java/com/example/myuniversityclient/ui/CateringFragment.kt
@@ -1,8 +1,6 @@
 package com.example.myuniversityclient.ui
 
 import android.content.Context
-import android.content.Intent
-import android.net.Uri
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
@@ -10,11 +8,10 @@ import android.view.ViewGroup
 import android.widget.Toast
 import androidx.fragment.app.Fragment
 import androidx.lifecycle.Observer
+import androidx.recyclerview.widget.DividerItemDecoration
 import androidx.recyclerview.widget.LinearLayoutManager
 import com.example.myuniversityclient.MainApplication
 import com.example.myuniversityclient.R
-import com.example.myuniversityclient.adapter.CateringHistoryAdapter
-import com.example.myuniversityclient.adapter.ServiceClickListener
 import com.example.myuniversityclient.data.models.CateringHistoryItem
 import com.example.myuniversityclient.data.models.CateringHistoryItemsList
 import com.example.myuniversityclient.domain.CateringFragmentViewModel
@@ -43,7 +40,8 @@ class CateringFragment : Fragment(), CateringHistoryAdapter.ServiceClickListener
 
         this.retainInstance = true
 
-        historyAdapter = CateringHistoryAdapter(history)
+        historyAdapter =
+            CateringHistoryAdapter(history)
         historyAdapter.listener = this
     }
 
@@ -55,8 +53,10 @@ class CateringFragment : Fragment(), CateringHistoryAdapter.ServiceClickListener
         val view: View = inflater.inflate(R.layout.fragment_catering, container, false)
 
         view.recyclerView.apply {
-            layoutManager = LinearLayoutManager(context)
+            val layoutManager = LinearLayoutManager(context)
             adapter = historyAdapter
+            this.layoutManager = layoutManager
+            addItemDecoration(DividerItemDecoration(context, layoutManager.orientation))
         }
 
         activity?.let {
