+++ app/src/main/res/layout/catering_history_item.xml
@@ -1,35 +1,34 @@
 <?xml version="1.0" encoding="utf-8"?>
-<LinearLayout
-    android:layout_width="match_parent"
+<LinearLayout android:layout_width="match_parent"
     android:layout_height="wrap_content"
+    xmlns:tools="http://schemas.android.com/tools"
     android:orientation="vertical"
     android:layout_gravity="center_horizontal"
-    android:padding="16dp"
+    android:padding="8dp"
     xmlns:android="http://schemas.android.com/apk/res/android">
 
     <TextView
         android:id="@+id/period"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:layout_gravity="left"
-        android:paddingLeft="16dp"
-        android:text="Title"
-        android:textStyle="bold"/>
+        android:layout_gravity="start"
+        tools:text="Title"
+        android:textAppearance="@style/TextAppearance.MaterialComponents.Body1"/>
 
     <TextView
         android:id="@+id/price"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:layout_gravity="left"
-        android:paddingLeft="16dp"
-        android:text="Description"/>
+        android:layout_gravity="start"
+        android:textAppearance="@style/TextAppearance.MaterialComponents.Body2"
+        tools:text="Description"/>
 
     <TextView
         android:id="@+id/items"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:layout_gravity="left"
-        android:paddingLeft="16dp"
-        android:text="Description"/>
+        android:layout_gravity="start"
+        android:textAppearance="@style/TextAppearance.MaterialComponents.Caption"
+        tools:text="Description"/>
 
 </LinearLayout>
