+++ app/src/main/res/layout/fragment_catering.xml
@@ -10,6 +10,6 @@
         android:id="@+id/recyclerView"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        android:layout_marginTop="24dp" />
+        tools:listitem="@layout/catering_history_item" />
 
 </FrameLayout>
