+++ app/src/main/res/layout/fragment_it_services.xml
@@ -11,7 +11,6 @@
         android:id="@+id/recyclerView"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        android:layout_marginTop="24dp"
         tools:listitem="@layout/title_subtitle_item"/>
 
 </FrameLayout>
